Cache cudf.Scalar (#11246)
Closes #11237 

This PR enables caching of `Scalar` objects. When the `Scalar()` constructor is invoked with the same arguments more than once, a cached instance is retrieved:

```python
>>> x1 = cudf.Scalar(1)
>>> x2 = cudf.Scalar(1)
>>> x1 is x2
True
```

It also adds registers a reinitialization hook with RMM that clears the cache. This ensures that no objects are left behind that hold on to the previous RMM memory resource after the call to `rmm.reinitialize()`.

Authors:
  - Ashwin Srinath (https://github.com/shwina)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/cudf/pull/11246