Special-case multibyte_split for single-byte delimiter (#11681)
multibyte_split does two scans to determine delimiters: One using the trie to determine matches, one to determine the offsets. For single-byte delimiters, the trie scan is unnecessary, since we can determine without context what is a delimiter. So I added a single `byte_split_kernel` by replacing the trie logic with a char comparison. The difference is quite significant:

| source_type | delim_size | delim_percent |    size_approx    |  GPU Time  | Peak Memory Usage | Encoded file size |
|-------------|------------|---------------|-------------------|------------|-------------------|-------------------|
|           0 |          1 |             1 | 2^30 = 1073741824 | 110.196 ms |         3.947 GiB |      1006.638 MiB |
|           0 |          2 |             1 | 2^30 = 1073741824 | 198.067 ms |         3.745 GiB |      1011.775 MiB |
|           0 |          1 |            25 | 2^30 = 1073741824 | 122.626 ms |         9.978 GiB |       762.462 MiB |
|           0 |          2 |            25 | 2^30 = 1073741824 | 224.000 ms |         6.163 GiB |       889.541 MiB |

This might point to the fact that the custom prefix scan implementation in multibyte_split is a bottleneck, but this needs more investigation.

Authors:
  - Tobias Ribizel (https://github.com/upsj)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Yunsong Wang (https://github.com/PointKernel)

URL: https://github.com/rapidsai/cudf/pull/11681