Disable very large column gtest for contiguous-split (#11706)
Disables a `ContiguousSplitUntypedTest` that simply creates a very large (over 3GB) column to test the output buffer size does not overflow. The gtests ends requiring 25GB of device memory when used with the arena allocator as mentioned in #11249. Very large columns like this should be not part of the unit test for libcudf.
This PR disables the test so it can be available for testing on specific conditions outside of CI.

Closes #11249

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/11706