Fix segmented-sort to ignore indices outside the offsets (#11888)
Fixes `cudf::segmented_sorted_order` to ignore indices outside the specified offsets values.

The segmented-sort function in general sorts subsets of the input using a column of offsets (integers) to identify the position of each segment. Here is an example:
```
input    = { 9, 8, 7, 6, 5, 4, 3, 2, 1, 0 }
offsets1 = { 0,       3,          7,      10 }
```
There are 3 segments to sort: `[0,3)`, `[3,7)`, and `[7,10)`
Segment 1 sorts to `{ 7, 8, 9 }`
Segment 2 sorts to `{ 3, 4, 5, 6 }`
Segment 3 sorts to `{ 0, 1, 2 }`
The segmented-sort result is `{ 7, 8, 9, 3, 4, 5, 6, 0, 1, 2 }`

If the offsets do not fully cover all the input the segmented-sort should ignore any segments outside of the offsets.
```
input    = { 9, 8, 7, 6, 5, 4, 3, 2, 1, 0 }
offsets2 = {          3,          7       }
```
Here there is only 1 segments to sort: `[3,7) => { 3, 4, 5, 6 }`
The segmented-sort result is `{ 9, 8, 7, 3, 4, 5, 6, 2, 1, 0 }`
The values before the first offset and after the last offset should be left unchanged.

The gtests have been corrected to expect this behavior. 
Also, the `SegmentedReductionTestUntyped.PartialSegmentReduction` gtest was improved to include offset gaps at the beginning and at the end to verify consistent behavior there as well.

Found while working on #11729

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - MithunR (https://github.com/mythrocks)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cudf/pull/11888