Fix cudf::stable_sorted_order for NaN and -NaN in FLOAT64 columns (#11874)
Fixes bug in `cudf::stable_sorted_order` when `-NaN` and `NaN` are in a FLOAT64 (double) columns.
The code was fixed by refactoring common code with `cudf::sorted_order`. This uses thrust sort functions to help align the behavior and keep results consistent.

New gtests were added to check for this case. Some test files were also updated per issue #11734 
The new tests are at the bottom of `sort_test.cpp` and `stable_sort_tests.cpp`

This was found while working on #11729 
The sorted-order functions are reused for many of the libcudf sort functions so this will help with the work in #11729

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/11874