Minor cleanup of root CMakeLists.txt for better organization (#11988)
Cleanup  some minor issues in the root cudf CMakeLists.txt. Make the seaching for `CUDA_SANITIZER` only occur when we are building tests as that doesn't need to be done for production builds. 

Move the gdb pretty print script logic to a separate region to better document what it is for.

Authors:
  - Robert Maynard (https://github.com/robertmaynard)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/11988