Add cudf::detail::sizes_to_offsets_iterator to allow checking overflow in offsets (#12180)
Add a new iterator that can be used with scan functions to also return the last element value with a higher precision than the scan type. This is used in the `cudf::strings::detail::make_strings_children` utility to convert output string sizes into offsets. The iterator used with `thrust::exclusive_scan` to compute an overall result that can be checked with the max of  `size_type`. The iterator provides minimal overhead to save the last entry of the scan. An error is thrown if the reduction value exceeds max of `size_type`.
A custom input iterator is not required since the `thrust::exclusive_scan` uses the init parameter type (set as 0) as the type used for the accumulator for the scan. The values are passed to the iterator with this type as well. The iterator then simply casts the output to the scan result iterator and saves the last value in a separate variable.

Closes #12087

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/12180