Use correct memory resource in io::make_column (#12364)
When creating a column from an `io::detail::column_buffer` object there is some subtlety in what memory resource should be used. Due to the way cudf's string APIs are designed, string columns in IO initially allocate temporary memory on construction of the `column_buffer` and only later allocate the final buffer to be returned to the user, while for all other types the initial allocation can be directly moved to the output. As a result, using the user-provided memory resource in the constructor of a `column_buffer` is sufficient to guarantee correct behavior for all column types except string columns, which also need a memory resource to be provided in `make_column`. This PR fixes the issue by storing the memory resource provided upon `column_buffer` construction, converting `make_column` into a method of the class, and removing the `mr` parameter to the method. With the new code, the saved `mr` is reused later ensuring that a consistent memory resource is used everywhere that needs to allocate memory returned to the user (as opposed to temporary memory that should be using `rmm::mr::get_current_device_resource()`).

Resolves #5823

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Divye Gala (https://github.com/divyegala)
  - Nghia Truong (https://github.com/ttnghia)
  - Yunsong Wang (https://github.com/PointKernel)

URL: https://github.com/rapidsai/cudf/pull/12364