Fix erroneously skipped ORC ZSTD test (#12486)
A C++ test that requires nvCOMP ZSTD is skipped when the required nvCOMP header is not present. This check does not seem to work correctly, as the test was skipped (on a local system) with nvCOMP 2.5.
This PR modifies the condition to skip the test to use more robust `is_decompression_disabled`.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Karthikeyan (https://github.com/karthikeyann)
  - https://github.com/nvdbaranec

URL: https://github.com/rapidsai/cudf/pull/12486