Update libcudf reduction docs for casting output types (#12526)
Fixes documentation on how `cudf::reduce` and `cudf::segmented_reduce` function when the input column type does not match the specified output data type. When the input and output types match, the reduction is performed using that type. When the types do not match, the input is promoted to either `double` or `int64_t` as appropriate through a transform-input-iterator when computing the reduction. The output is then cast to the specified output data type. This works for unsigned  values as well. For UINT64 types, this works for all values where the result would be <= max(int64_t). For values where the result may be > max(int64_t) the `output_dtype == UINT64` and falls into the case where the types match and reduction is performed on that type.
This PR changes only the comments and doxygen for the reduce APIs.

Closes #9988

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/12526