Pin cuda-nvrtc. (#12606)
Ensure we pin `cuda-nvrtc` the same way as `cuda-nvrtc-dev`. This is necessary to avoid pulling a newer `cuda-nvrtc` at build time. The `-dev` package is pinned to `cuda-nvrtc >= (dev version)` but we want the pinnings of both packages to be the same, and not let `cuda-nvrtc` float ahead.

Authors:
  - Bradley Dice (https://github.com/bdice)
  - David Wendt (https://github.com/davidwendt)
  - AJ Schmidt (https://github.com/ajschmidt8)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cudf/pull/12606