Fix an issue with parquet chunked reader undercounting string lengths. (#12859)
Fixes https://github.com/NVIDIA/spark-rapids/issues/7796

The way we were counting string lengths was causing some values to be discarded in some cases.  Essentially, dictionary indices are encoded in runs.  When decoding to a particular position, the code occasionally has to decode slightly past that.  The code that stores the indices buffers these extra values across calls, but the code that was adding up the lengths was just ignoring them.

The solution is to simply process all the string lengths in one batch (which also opens up the door to using a seperate warp to do this work in parallel with the rest of the level decoding happening at the time).

Also includes a minor optimization in the call to gpuInitStringDescriptors - a template parameter that controls whether or not we write out dictionary indices to the rolling buffer - a step which is unnecessary during preprocessing but we were doing anyway.

Authors:
  - https://github.com/nvdbaranec
  - Nghia Truong (https://github.com/ttnghia)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/12859