Remove return type from @return doxygen tags (#12908)
The doxygen `@return` does not require a type but only the description. However, the doxygen output looks cleaner without the type. There were only a few places in the public header files that needed to be corrected.
This technically a documentation-only change.
Found when working on #12904

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Karthikeyan (https://github.com/karthikeyann)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/12908