Declare a different name for nan_equality.UNEQUAL to prevent Cython warnings. (#12947)
Closes #9462.

```
warning: cudf/_lib/cpp/types.pxd:51:8: 'UNEQUAL' redeclared
```

We've been dealing with warnings like this for as long as I've worked on cudf. I think it'd be good to fix this by renaming the Cython enum. We don't rely on this name very much at the Cython/Python level, so it's a small change. (This has been irking me for a long time and I want to see this warning go away in the build logs.)

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - https://github.com/brandon-b-miller

URL: https://github.com/rapidsai/cudf/pull/12947