Set the null count in output columns in the CSV reader (#13221)
CSV reader used to keep the null count of the output columns as `UNKNOWN_NULL_COUNT`. This has performance implications on further use of these columns.
This PR adds a validity counter to convert_csv_to_cudf kernel. This counter is then used to set the correct null count in the columns.
Also added a test that explicitly checks the null counts, as it is never explicitly used in the CSV reader tests.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/13221